[modularize] Reconstruct codes into `serialization` module.
